requirements.txt
alembic==1.15.2
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
et_xmlfile==2.0.0
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fsspec==2025.3.0
future==1.0.0
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
llvmlite==0.44.0
Mako==1.3.9
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
openai-whisper==20240930
openpyxl==3.1.5
pandas==2.2.3
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.1
regex==2024.11.6
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.39
srt==3.5.3
sympy==1.13.3
tiktoken==0.9.0
torch==2.2.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vosk==0.3.44
websockets==15.0.1
Werkzeug==3.1.3
